First shot at pushing GH update from Jenkins
