+++ scripts/node-jenkins-status.js
@@ -1,17 +1,13 @@
 'use strict'
 
-const pollJenkins = require('../lib/poll-jenkins')
+const pushJenkinsUpdate = require('../lib/push-jenkins-update')
 
 module.exports = function (app) {
-  // to trigger polling manually
-  app.get('/jenkins/:prId', (req, res) => {
-    const prId = req.params.prId
-
-    pollJenkins({
-      owner: 'nodejs',
-      repoName: 'node',
-      prId: parseInt(prId, 10)
-    })
+  app.post('/node/jenkins', (req, res) => {
+    pushJenkinsUpdate({
+      owner: 'TestOrgPleaseIgnore',
+      repoName: 'node'
+    }, req.body)
 
     res.end()
   })
