+++ test/_fixtures/error-payload.json
@@ -1,0 +1,12 @@
+{
+    "fullDisplayName": "node-test-commit-arm #3087",
+    "result": "FAILURE",
+    "changeSet": {
+        "items": [
+            {
+                "commitId": "7190bd23810f0880d95605bf6f785224a345d477"
+            }
+        ]
+    },
+    "url": "https://ci.nodejs.org/job/node-test-commit-arm/3087/"
+}
