+++ test/_fixtures/pending-payload.json
@@ -1,0 +1,12 @@
+{
+    "fullDisplayName": "node-test-commit-linux #3176",
+    "result": null,
+    "changeSet": {
+        "items": [
+            {
+                "commitId": "7190bd23810f0880d95605bf6f785224a345d477"
+            }
+        ]
+    },
+    "url": "https://ci.nodejs.org/job/node-test-commit-linux/3176/"
+}
